backend/requirements.txt
# FastAPI ve server
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# HTTP Client
httpx==0.27.2
aiohttp==3.11.2

# OpenTelemetry - Tracing ve Monitoring
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-instrumentation-httpx==0.49b2
opentelemetry-instrumentation-sqlalchemy==0.49b2
opentelemetry-exporter-otlp==1.28.2
# opentelemetry-exporter-jaeger - Deprecated, OTLP kullanıyoruz

# CORS
python-multipart==0.0.12

# Utilities
python-dotenv==1.0.1